+++ tests/end2end/test_zhmc_partition.py
@@ -23,9 +23,11 @@ import pytest
 import mock
 import random
 import requests.packages.urllib3
-from ansible.module_utils import six
-import zhmcclient
-from zhmcclient.testutils.hmc_definition_fixtures import hmc_definition, hmc_session  # noqa: F401, E501
+# pylint: disable=line-too-long,unused-import
+from zhmcclient.testutils import hmc_definition, hmc_session  # noqa: F401, E501
+from zhmcclient.testutils import dpm_mode_cpcs  # noqa: F401, E501
+# pylint: enable=line-too-long,unused-import
+
 from plugins.modules import zhmc_partition
 from .utils import mock_ansible_module, get_failure_msg
 
@@ -66,11 +68,15 @@ def get_module_output(mod_obj):
     return func(*call_args[0], **call_args[1])
 
 
-def assert_partition_props(partition_props):
+def assert_partition_props(partition_props, where):
     """
     Assert the output object of the zhmc_partition module
     """
-    assert isinstance(partition_props, dict)  # Dict of Partition properties
+
+    # Assert partition_props (dict of Partition properties)
+    assert isinstance(partition_props, dict), \
+        "{w}: partition_props has invalid type {t}". \
+        format(w=where, t=type(partition_props))
 
     # Assert presence of normal properties in the output
     for prop_name in zhmc_partition.ZHMC_PARTITION_PROPERTIES:
@@ -79,97 +85,105 @@ def assert_partition_props(partition_props):
             # TODO: Do better than just skipping them - they are present under
             # certain conditions.
             continue
-        assert prop_name_hmc in partition_props
+        assert prop_name_hmc in partition_props, \
+            "{w}: partition_props misses property {pn}: {pp}". \
+            format(w=where, pn=prop_name_hmc, pp=partition_props)
 
     # Assert presence of the artificial properties in the output
 
-    assert 'hbas' in partition_props
+    assert 'hbas' in partition_props, where
     hba_props_list = partition_props['hbas']
     if hba_props_list is not None:
-        assert isinstance(hba_props_list, list)  # List of HBAs
+        assert isinstance(hba_props_list, list), where  # List of HBAs
         for hba_props in hba_props_list:
-            assert isinstance(hba_props, dict)  # Dict of HBA properties
+            assert isinstance(hba_props, dict), where  # Dict of HBA properties
 
-    assert 'nics' in partition_props
+    assert 'nics' in partition_props, where
     nic_props_list = partition_props['nics']
     if nic_props_list is not None:
-        assert isinstance(nic_props_list, list)  # List of NICs
+        assert isinstance(nic_props_list, list), where  # List of NICs
         for nic_props in nic_props_list:
-            assert isinstance(nic_props, dict)  # Dict of NIC properties
+            assert isinstance(nic_props, dict), where  # Dict of NIC properties
 
-    assert 'virtual-functions' in partition_props
+    assert 'virtual-functions' in partition_props, where
     vf_props_list = partition_props['virtual-functions']
     if vf_props_list is not None:
-        assert isinstance(vf_props_list, list)  # List of VFs
+        assert isinstance(vf_props_list, list), where  # List of VFs
         for vf_props in vf_props_list:
-            assert isinstance(vf_props, dict)  # Dict of VF properties
+            assert isinstance(vf_props, dict), where  # Dict of VF properties
 
 
 @pytest.mark.parametrize(
-    "check_mode", [False, True])
+    "check_mode", [
+        pytest.param(False, id="check_mode=False"),
+        pytest.param(True, id="check_mode=True"),
+    ]
+)
 @pytest.mark.parametrize(
-    "partition_type", ['ssc', 'linux'])
+    "partition_type", [
+        pytest.param('ssc', id="partition_type=ssc"),
+        pytest.param('linux', id="partition_type=linux"),
+    ]
+)
 @mock.patch("plugins.modules.zhmc_partition.AnsibleModule", autospec=True)
 def test_partition_facts(
-        ansible_mod_cls, partition_type, check_mode, hmc_session):  # noqa: F811, E501
+        ansible_mod_cls, partition_type, check_mode, dpm_mode_cpcs):  # noqa: F811, E501
     """
     Test fact gathering on a partition.
     """
+    if not dpm_mode_cpcs:
+        pytest.skip("HMC definition does not include any CPCs in DPM mode")
 
-    hd = hmc_session.hmc_definition
+    for cpc in dpm_mode_cpcs:
+        assert cpc.dpm_enabled
 
-    hmc_auth = dict(userid=hd.hmc_userid, password=hd.hmc_password)
-    if isinstance(hd.hmc_verify_cert, six.string_types):
-        hmc_auth['ca_certs'] = hd.hmc_verify_cert
-    elif isinstance(hd.hmc_verify_cert, bool):
-        hmc_auth['verify'] = hd.hmc_verify_cert
+        session = cpc.manager.session
+        hd = session.hmc_definition
+        hmc_host = hd.host
+        hmc_auth = dict(userid=hd.userid, password=hd.password,
+                        ca_certs=hd.ca_certs, verify=hd.verify)
 
-    # Determine one of the CPCs in the HMC definition file to test
-    cpc_names = list(hd.cpcs.keys())
-    assert len(cpc_names) >= 1
-    cpc_name = cpc_names[0]
+        # Determine a random existing partition of the desired type to test.
+        partitions = cpc.partitions.list()
+        typed_partitions = [p for p in partitions
+                            if p.get_property('type') == partition_type]
+        if len(typed_partitions) == 0:
+            pytest.skip("CPC '{c}' has no partitions of type '{t}'".
+                        format(c=cpc.name, t=partition_type))
+        partition = random.choice(typed_partitions)
 
-    client = zhmcclient.Client(hmc_session)
-    cpc = client.cpcs.find_by_name(cpc_name)
+        where = "partition '{p}'".format(p=partition.name)
 
-    # Determine a random existing partition of the desired type to test.
-    partitions = cpc.partitions.list()
-    assert len(partitions) >= 1
-    typed_partitions = [p for p in partitions
-                        if p.get_property('type') == partition_type]
-    if len(typed_partitions) == 0:
-        pytest.skip("CPC '{c}' has no partitions of type '{t}'".
-                    format(c=cpc_name, t=partition_type))
-    partition = random.choice(typed_partitions)
+        # Prepare module input parameters (must be all required + optional)
+        params = {
+            'hmc_host': hmc_host,
+            'hmc_auth': hmc_auth,
+            'cpc_name': cpc.name,
+            'name': partition.name,
+            'state': 'facts',
+            'properties': {},
+            'expand_storage_groups': False,
+            'expand_crypto_adapters': False,
+            'log_file': None,
+            '_faked_session': None,
+        }
 
-    # Prepare module input parameters (must be all required + optional)
-    params = {
-        'hmc_host': hd.hmc_host,
-        'hmc_auth': hmc_auth,
-        'cpc_name': cpc_name,
-        'name': partition.name,
-        'state': 'facts',
-        'properties': {},
-        'expand_storage_groups': False,
-        'expand_crypto_adapters': False,
-        'log_file': None,
-        '_faked_session': None,
-    }
+        # Prepare mocks for AnsibleModule object
+        mod_obj = mock_ansible_module(ansible_mod_cls, params, check_mode)
 
-    # Prepare mocks for AnsibleModule object
-    mod_obj = mock_ansible_module(ansible_mod_cls, params, check_mode)
+        # Exercise the code to be tested
+        with pytest.raises(SystemExit) as exc_info:
+            zhmc_partition.main()
+        exit_code = exc_info.value.args[0]
 
-    # Exercise the code to be tested
-    with pytest.raises(SystemExit) as exc_info:
-        zhmc_partition.main()
-    exit_code = exc_info.value.args[0]
+        # Assert module exit code
+        assert exit_code == 0, \
+            "{w}: Module failed with exit code {e} and message:\n{m}". \
+            format(w=where, e=exit_code, m=get_failure_msg(mod_obj))
 
-    # Assert module exit code
-    assert exit_code == 0, \
-        "Module unexpectedly failed with this message:\n{0}". \
-        format(get_failure_msg(mod_obj))
-
-    # Assert module output
-    changed, partition_props = get_module_output(mod_obj)
-    assert changed is False
-    assert_partition_props(partition_props)
+        # Assert module output
+        changed, partition_props = get_module_output(mod_obj)
+        assert changed is False, \
+            "{w}: Module returned changed={c}". \
+            format(w=where, c=changed)
+        assert_partition_props(partition_props, where)